Fixes bug where last ramp would be shortcutted forever.
Due to numerical precision issues, it was possible for the TryBlend function to attempt a blend on the last ramp that did not include the last ramp, which led to the last ramp of a trajectory being shortcutted forever.  This fix extends the range of blends by EpsilonT, the configured time epsilon, to ensure that shortcutted ramps are included in that range.

It also cleans up some minor whitespace and reuses the EpsilonV argument where appropriate.